combined merge_hex
